fix: extract components themes into multiple files
