+++ admin/src/app/mui_joy/themes/checkboxTheme.js
@@ -1,0 +1,22 @@
+const checkboxTheme = {
+	JoyCheckbox: {
+		styleOverrides: {
+			root: ( { ownerState } ) => ( {
+				wordBreak: 'break-word', // handle break of long urls that are as checkbox labels
+				...( ownerState.ellipsis === true ? {
+					overflow: 'hidden',
+				} : null ),
+			} ),
+			label: ( { ownerState } ) => ( {
+				wordBreak: 'break-word', // handle break of long urls that are as checkbox labels
+				...( ownerState.ellipsis === true ? {
+					textOverflow: 'ellipsis',
+					overflow: 'hidden',
+					whiteSpace: 'nowrap',
+				} : null ),
+			} ),
+		},
+	},
+};
+
+export default checkboxTheme;
